# kapsule-incus-manager/api/schema/openapi.yaml
openapi: "3.1.0"
info:
  title: Kapsule-Incus-Manager API
  version: "0.1.0"
  description: >
    Canonical API definition for the KIM daemon.
    All three frontends (QML, web, CLI) are thin clients against this API.
    The D-Bus interface mirrors every operation defined here.
  license:
    name: GPL-3.0-or-later
servers:
  - url: http://127.0.0.1:8765
    description: Local KIM daemon
...
components:
  schemas:
...
    Operation:
      type: object
      required: [id, description, status, created_at]
      properties:
        id:          { type: string, format: uuid }
        description: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
        metadata:    { type: object, additionalProperties: true }
        error:       { type: string }
    InstanceStatus:
      type: string
      enum: [Running, Stopped, Frozen, Error, Unknown]
    InstanceType:
      type: string
      enum: [container, virtual-machine]
    InstanceSummary:
      type: object
      required: [name, type, status, project, remote]
      properties:
        name:       { type: string }
        type:       { $ref: "#/components/schemas/InstanceType" }
        status:     { $ref: "#/components/schemas/InstanceStatus" }
        image:      { type: string }
        project:    { type: string }
        remote:     { type: string }
        created_at: { type: string, format: date-time }
        cpu_usage:  { type: number, description: "CPU usage 0.0–1.0" }
        memory_usage_bytes: { type: integer }
        disk_usage_bytes:   { type: integer }
...
  responses:
...
    Accepted:
      description: Operation accepted and running asynchronously
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Operation" }
...
paths:
...
  /api/v1/provisioning/waydroid:
    post:
      operationId: createWaydroidContainer
      summary: Provision an Incus container with Waydroid pre-installed
      tags: [provisioning/waydroid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                image:       { type: string, default: "images:ubuntu/22.04/cloud" }
                image_type:  { type: string, enum: [VANILLA, GAPPS], default: VANILLA }
                arch:        { type: string, enum: [x86_64, arm64], default: x86_64 }
                gpu:         { type: boolean, default: false }
                gpu_type:    { type: string, default: physical }
                profiles:    { type: array, items: { type: string } }
                disk_size:   { type: string, default: "40GB" }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/waydroid/{name}/extensions:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listWaydroidExtensions
      summary: List installed Waydroid extensions
      tags: [provisioning/waydroid]
      responses:
        "200":
          description: Extension list
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    post:
      operationId: installWaydroidExtension
      summary: Install a Waydroid extension
      tags: [provisioning/waydroid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [extension]
              properties:
                extension: { type: string, enum: [gapps, microg, magisk, libhoudini, libndk, widevine, keymapper] }
                project:   { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/waydroid/{name}/extensions/{extension}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: extension, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: removeWaydroidExtension
      summary: Remove a Waydroid extension
      tags: [provisioning/waydroid]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/waydroid/{name}/backups:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listWaydroidBackups
      summary: List Waydroid backups inside the container
      tags: [provisioning/waydroid]
      responses:
        "200":
          description: Backup list
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    post:
      operationId: backupWaydroid
      summary: Create a Waydroid data backup
      tags: [provisioning/waydroid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dest:    { type: string }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/waydroid/{name}/restore:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: restoreWaydroid
      summary: Restore a Waydroid backup
      tags: [provisioning/waydroid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [archive]
              properties:
                archive: { type: string }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/waydroid/{name}/cloud-sync:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    put:
      operationId: configureWaydroidCloudSync
      summary: Configure rclone cloud sync for Waydroid backups
      tags: [provisioning/waydroid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [remote_name, remote_path]
              properties:
                remote_name: { type: string }
                remote_path: { type: string }
                schedule:    { type: string }
                project:     { type: string, default: "" }
      responses:
        "200":
          description: Cloud sync configured
  /api/v1/provisioning/waydroid/{name}/gpus:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: attachWaydroidGpu
      summary: Attach a GPU to a Waydroid container
      tags: [provisioning/waydroid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dev_name: { type: string, default: "gpu0" }
                gpu_type: { type: string, default: physical }
                pci:      { type: string }
                vendor:   { type: string }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/waydroid/{name}/gpus/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: detachWaydroidGpu
      summary: Detach a GPU from a Waydroid container
      tags: [provisioning/waydroid]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/waydroid/fleet:
    get:
      operationId: listWaydroidFleet
      summary: List Waydroid containers
      tags: [provisioning/waydroid]
      parameters:
        - { name: project, in: query, schema: { type: string, default: "" } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Container list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InstanceSummary" } }
  /api/v1/provisioning/waydroid/publish:
    post:
      operationId: publishWaydroidContainer
      summary: Publish a Waydroid container as a reusable Incus image
      tags: [provisioning/waydroid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                alias:       { type: string }
                description: { type: string }
                public:      { type: boolean, default: false }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }

  # ── macOS VM provisioning ─────────────────────────────────────────────────
...